requirements.txt
fastapi
uvicorn
pymongo
pydantic
pyjwt
passlib
python-dotenv
certifi
python-jose
bcrypt